Tilføjet metode til at check efter admin rolle, og sørget for at admin kan oprette en film og gemme den, og knapper

+++ src/main/java/com/kinoxp/controller/MovieController.java
@@ -3,22 +3,33 @@ package com.kinoxp.controller;
 import com.kinoxp.dto.MovieRequest;
 import com.kinoxp.dto.MovieResponse;
 import com.kinoxp.model.movie.*;
+import com.kinoxp.model.user.Role;
+import com.kinoxp.model.user.User;
+import com.kinoxp.repository.UserRepository;
+import com.kinoxp.security.AdminChecker;
 import com.kinoxp.service.MovieService;
+import com.kinoxp.service.UserService;
 import jakarta.validation.Valid;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/kino/movies")
 public class MovieController {
 
     private final MovieService movieService;
+    private final UserService userService;
+    private final UserRepository userRepository;
 
-    public MovieController(MovieService movieService) {
+    public MovieController(MovieService movieService, UserService userService, UserRepository userRepository) {
         this.movieService = movieService;
+        this.userService = userService;
+        this.userRepository = userRepository;
     }
 
     @GetMapping
@@ -49,21 +60,33 @@ public class MovieController {
     }
 
     @PostMapping
-    public ResponseEntity<MovieResponse> createMovie(@Valid @RequestBody MovieRequest request) {
+    public ResponseEntity<MovieResponse> createMovie(@Valid @RequestBody MovieRequest request, @RequestParam Long userId) {
+        if (!AdminChecker.isAdmin(userService, userId)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+    }
+
         MovieResponse createdMovie = movieService.createMovie(request);
         URI location = URI.create("/kino/movies/" + createdMovie.movieId());
         return ResponseEntity.created(location).body(createdMovie);
     }
 
     @PutMapping("/{movieId}")
-    public ResponseEntity<MovieResponse> updateMovie(@PathVariable Long movieId, @Valid @RequestBody MovieRequest request) {
+    public ResponseEntity<MovieResponse> updateMovie(@PathVariable Long movieId, @Valid @RequestBody MovieRequest request, @RequestParam Long userId) {
+        if (!AdminChecker.isAdmin(userService, userId)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         return movieService.updateMovie(movieId, request)
                 .map(ResponseEntity::ok)
                 .orElseGet(() -> ResponseEntity.notFound().build());
     }
 
     @DeleteMapping("/{movieId}")
-    public ResponseEntity<Void> deleteMovie(@PathVariable Long movieId) {
+    public ResponseEntity<Void> deleteMovie(@PathVariable Long movieId, @RequestParam Long userId) {
+        if(!AdminChecker.isAdmin(userService, userId)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         boolean deleted = movieService.deleteMovieById(movieId);
         if (!deleted) return ResponseEntity.notFound().build();
